/tftpboot -> ${X_TFTPBOOT}
I'd like to eventually support overriding tftpboot per-configuration
file and globally so we can do a separate install dir per config file.
(Eg, if we're building download image sets.)

Whilst we're here, cp -f somewhere else.

Also whilst we're here, comment a hack that I merged in that
populates a single image for carambola2.  It should be a build
target, not a hack in build_uboot.